Print summary for free energy profile

`summary.profiles` prints the list of free energy minima with maximal and minimal free energy differences.

# S3 method for profiles
summary(object, imind = 1, imaxd = NULL, ...)

Arguments

object

profiles object.

imind

index of a hill from which calculation of difference starts (default 1).

imaxd

index of a hill from which calculation of difference stops (default the rest of hills).

...

further arguments passed to or from other methods.

Examples

tfes<-fes(acealanme, imax=5000)
minima<-fesminima(tfes)
prof<-feprof(minima)
summary(prof)
#>   letter CV1bin CV2bin        CV1        CV2 free_energy   min diff  max diff
#> 1      A     79    235 -1.2196771  2.6241539   -58.22765  0.0000000  0.000000
#> 2      B     28    242 -2.4763142  2.7966335   -55.37985 -2.1604229  3.885774
#> 3      C     70    121 -1.4414366 -0.1847996   -54.65751  0.1928119 12.342623
#> 4      D    166    153  0.9239978  0.6036786   -54.17444  0.6563058 21.574352
#> 5      E     30    134 -2.4270343  0.1355197   -52.74630  0.7541017 12.342623
#> 6      F    172    249  1.0718375  2.9691131   -46.45687  0.7541017 21.574351
#>        tail
#> 1  0.000000
#> 2  1.676195
#> 3  2.789169
#> 4  5.485229
#> 5  5.267991
#> 6 12.951382